extract-maybe-register-hicache-draft-move: Move maybe_register_hicache_draft to mem_cache.kv_cache_builder
Mechanical cut + paste for the ``extract-maybe-register-hicache-draft``
mech move.

Cut ``Scheduler.maybe_register_hicache_draft`` (@staticmethod after the
prep commit) and append to ``mem_cache/kv_cache_builder.py``. Drop
``@staticmethod`` decorator; dedent body to module level. Collapse
self-module qualifier ``kv_cache_builder.get_draft_kv_pool(...)`` → bare
``get_draft_kv_pool(...)`` since both functions now live in the same
module.

Add a module logger to ``kv_cache.py`` (the moved function calls
``logger.warning(...)``).

Sole caller in ``Scheduler.__init__`` updated from
``Scheduler.maybe_register_hicache_draft(...)`` →
``kv_cache_builder.maybe_register_hicache_draft(...)`` (pure prefix replacement).

Refactor chain ID: extract-maybe-register-hicache-draft-move
